feat: criacao das views para usuario, removendo o swagger,

+++ TaskFlow/Controllers/UserController.cs
@@ -4,8 +4,6 @@ using TaskFlow.Domain.Entity;
 
 namespace TaskFlow.Controllers
 {
-    [ApiController]
-    [Route("api/v1/user")]
     public class UserController : Controller
     {
         private readonly IUnitOfWorkService _serviceUoW;
@@ -15,46 +13,79 @@ namespace TaskFlow.Controllers
             _serviceUoW = unitOfWorkService;
         }
 
+        [HttpGet]
+        public IActionResult Create()
+        {
+            return View();
+        }
+
         [HttpPost]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<IActionResult> AddUser([FromBody] UserEntity userEntity)
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Create(UserEntity userEntity)
         {
             if (!ModelState.IsValid)
-                return BadRequest(ModelState);
+                return View(userEntity);
 
             var result = await _serviceUoW.UserService.AddUserAsync(userEntity);
-            return result.Success ? Ok(result) : BadRequest(result);
+            if (result.Success)
+                return RedirectToAction("Success");
+
+            ModelState.AddModelError("", "Erro ao cadastrar usuário.");
+            return View(userEntity);
         }
 
-        [HttpPut]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesDefaultResponseType]
-        public async Task<IActionResult> UpdateUser([FromBody] UserEntity userEntity)
-        {
-            var result = await _serviceUoW.UserService.UpdateUserAsync(userEntity);
-            return result.Success ? Ok(result) : BadRequest(userEntity);
-        }
-
-        [HttpDelete("{id}")]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesDefaultResponseType]
-        public async Task<IActionResult> DeleteUser(int id)
-        {
-            await _serviceUoW.UserService.DeleteUserAsync(id);
-            return Ok();
-        }
-
-        [HttpGet("All")]
-        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(List<UserEntity>))]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> GetAllUsers()
+        [HttpGet]
+        public async Task<IActionResult> List()
         {
             var users = await _serviceUoW.UserService.GetAllUsersAsync();
-            return Ok(users);
+            return View(users);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> Delete(int id)
+        {
+            if (id <= 0)
+                return BadRequest("ID inválido.");
+
+            await _serviceUoW.UserService.DeleteUserAsync(id);
+
+            ModelState.AddModelError("", "Erro ao excluir usuário.");
+            return RedirectToAction("List");
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Edit(UserEntity userEntity)
+        {
+            if (!ModelState.IsValid)
+                return View(userEntity);
+
+            var result = await _serviceUoW.UserService.UpdateUserAsync(userEntity);
+
+            if (result.Success)
+                return RedirectToAction("List");
+
+            ModelState.AddModelError("", "Erro ao editar usuário.");
+            return View(userEntity);
+        }
+
+        [HttpGet]
+        public async Task<IActionResult> Edit(int id)
+        {
+            if (id <= 0)
+                return BadRequest("ID inválido.");
+
+            var user = await _serviceUoW.UserService.GetUserByIdAsync(id);
+
+            if (user == null)
+                return NotFound();
+
+            return View(user);
+        }
+
+        public IActionResult Success()
+        {
+            return View();
         }
     }
 }
